+++ src/markdown/directives/yaml.js
@@ -7,13 +7,13 @@ import directives from './directives'
  *
  * @alias module:markdown/directives/yaml
  * @param {String} text Target text.
- * @param {boolean} allowLazy By `true`, it try to parse lazy YAML in defined
+ * @param {boolean} allowLoose By `true`, it try to parse loose YAML in defined
  *     directives.
  * @returns {Object|false} Return parse result, or `false` when failed to parse.
  */
 
 const keyPattern = `[_$]?(?:${directives.join('|')})`
-const lazyMatcher = new RegExp(`^(${keyPattern}\\s*:)(.+)$`)
+const looseMatcher = new RegExp(`^(${keyPattern}\\s*:)(.+)$`)
 const specialChars = `["'{|>~&*`
 
 function parse(text) {
@@ -27,12 +27,12 @@ function parse(text) {
   }
 }
 
-function convertLazy(text) {
+function convertLoose(text) {
   return text
     .split(/\r?\n/)
     .reduce(
       (ret, line) =>
-        `${ret}${line.replace(lazyMatcher, (original, prop, value) => {
+        `${ret}${line.replace(looseMatcher, (original, prop, value) => {
           const trimmed = value.trim()
 
           if (trimmed.length === 0 || specialChars.includes(trimmed[0]))
@@ -48,4 +48,5 @@ function convertLazy(text) {
     .trim()
 }
 
-export default (text, allowLazy) => parse(allowLazy ? convertLazy(text) : text)
+export default (text, allowLoose) =>
+  parse(allowLoose ? convertLoose(text) : text)