+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Breaking
+
+- Renamed `lazyYAML` constructor option into `looseYAML`. ([#68](https://github.com/marp-team/marpit/pull/68))
+
 ### Changed
 
 - Move documentation from README.md to [https://marpit.marp.app/](https://marpit.marp.app/) ([#67](https://github.com/marp-team/marpit/pull/67))